+++ js/button/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @gradio/button
 
+## 0.2.45
+
+### Dependency updates
+
+- @gradio/upload@0.11.4
+- @gradio/client@1.2.1
+
+## 0.2.44
+
+### Dependency updates
+
+- @gradio/utils@0.5.0
+- @gradio/client@1.2.0
+- @gradio/upload@0.11.3
+
 ## 0.2.43
 
 ### Dependency updates