+++ js/downloadbutton/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @gradio/downloadbutton
 
+## 0.1.22
+
+### Dependency updates
+
+- @gradio/client@1.2.1
+- @gradio/button@0.2.45
+
+## 0.1.21
+
+### Dependency updates
+
+- @gradio/utils@0.5.0
+- @gradio/client@1.2.0
+- @gradio/button@0.2.44
+
 ## 0.1.20
 
 ### Dependency updates