+++ js/lite/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @gradio/lite
 
+## 4.37.2
+
+### Dependency updates
+
+- gradio@4.37.2
+
+## 4.37.1
+
+### Dependency updates
+
+- gradio@4.37.1
+
+## 4.37.0
+
+### Dependency updates
+
+- @gradio/wasm@0.11.0
+- gradio@4.37.0
+
 ## 4.36.1
 
 ### Dependency updates