+++ js/uploadbutton/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @gradio/uploadbutton
 
+## 0.6.13
+
+### Dependency updates
+
+- @gradio/upload@0.11.4
+- @gradio/client@1.2.1
+- @gradio/button@0.2.45
+
+## 0.6.12
+
+### Dependency updates
+
+- @gradio/utils@0.5.0
+- @gradio/client@1.2.0
+- @gradio/button@0.2.44
+- @gradio/upload@0.11.3
+
 ## 0.6.11
 
 ### Dependency updates